More Bug Fixes
o Fixes to ensure that the transmitter IDs with a callsign at the end of its first transmission
o Fixes to sleep at the end of an event
o Fixes for crashes and incorrect behavior when setting and launching events.
o Run button is disabled when an event is disabled, and when writing event changes to file.